requirements.txt
asgiref==3.8.1
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
blinker==1.8.2
bracex==2.5
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cssselect==1.2.0
diff-match-patch==20230430
Django==4.2.15
django-cors-headers==4.4.0
django-filter==24.3
django-framework==0.1.2
django-import-export==4.1.1
django-jazzmin==3.0.0
django-shortuuidfield==0.1.3
djangorestframework==3.15.2
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.7.1
flask==3.0.3
ghp-import==2.1.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
importlib-metadata==8.2.0
inflection==0.5.1
itsdangerous==2.2.0
jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lxml==5.3.0
Markdown==3.6
markdown2==2.5.0
MarkupSafe==2.1.5
mergedeep==1.3.4
packaging==24.1
paginate==0.5.6
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.2.2
pygments==2.18.0
pymdown-extensions==10.9
pyquery==2.0.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
pyyaml-env-tag==0.1
readtime==3.0.0
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rpds-py==0.20.0
shortuuid==1.0.13
six==1.16.0
smmap==5.0.1
soupsieve==2.6
sqlparse==0.5.1
tablib==3.5.0
typing-extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
watchdog==4.0.2
wcmatch==9.0
werkzeug==3.0.3
zipp==3.19.2
